requirements.txt
tornado
pynsq
futures
mdict